requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
click==8.2.1
    # via uvicorn
dnspython==2.8.0
    # via pymongo
fastapi==0.116.1
    # via -r requirements.in
graphql-core==3.2.6
    # via strawberry-graphql
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
lia-web==0.2.3
    # via strawberry-graphql
packaging==25.0
    # via strawberry-graphql
pydantic==2.11.9
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pymongo==4.9.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.1.1
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via
    #   -r requirements.in
    #   fastapi
strawberry-graphql==0.282.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   lia-web
    #   pydantic
    #   pydantic-core
    #   strawberry-graphql
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
